feat&fix: add online mode and fix score board
