+++ data/score.txt
@@ -8,3 +8,9 @@ asdasd 50 3521
 asdasd 50 3569
 asdasd 60 2649
 asdasd 50 2408
+urikpro64 800 14721
+asdasd 1720 56386
+dfsdf 1780 55394
+asda 50 2168
+sad 80 4073
+asdasd 60 3870
